requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.10
mutagen==1.47.0
pathvalidate==3.2.1
pick==1.6.0
python-dotenv==1.0.1
qobuz==0.0.2
qobuz-dl==0.9.9.10
requests==2.32.3
schedule==1.2.2
soupsieve==2.6
tqdm==4.66.5
urllib3==2.2.3
